Suppress warning s404
