+++ src/ssb_project_cli/ssb_project/__init__.py
@@ -1,1 +1,1 @@
-"""ssb-project."""
+"""ssb-project package."""
