+++ src/ssb_project_cli/ssb_project/clean/__init__.py
@@ -1,0 +1,1 @@
+"""Clean command package."""
