+++ src/ssb_project_cli/ssb_project/create/__init__.py
@@ -1,0 +1,1 @@
+"""Create command package."""
