+++ src/ssb_project_cli/ssb_project/create/repo_privacy.py
@@ -1,0 +1,10 @@
+"""This module contains the RepoPrivacy class."""
+from enum import Enum
+
+
+class RepoPrivacy(str, Enum):
+    """Class with predefined privacy enums."""
+
+    internal = "internal"
+    private = "private"
+    public = "public"
