+++ tests/clean_test/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for clean package."""
